Crux.Rest.ApiError exception (crux_rest v0.3.0-dev)

Represents a Discord API error.

Raised or returned whenever the api responded with a non 2xx status code.

Types

t()
(since 0.1.0)

Specs

t() :: %Crux.Rest.ApiError{
  __exception__: true,
  code: integer() | nil,
  message: String.t(),
  method: Crux.Rest.Request.method(),
  path: String.t(),
  status_code: integer()
}

Field	Description	Example(s)

status_code	The HTTP Response code	400, 404, 403

code	Discord's JSON Error Code	10006, 90001

message	Message describing the error	Unknown Invite

path	Path of the request	/invites/broken

method	HTTP verb of the request	:get, :post, :patch

In case an error response was sent by CloudFlare, code will be nil and message a HTML document describing the error.

Functions

exception(msg)

Specs

exception(message :: binary()) :: Exception.t()

Default implementation only providing a message for raise/2.

Not internally used.

exception(request, response)
(since 0.1.0)

Specs

exception(Crux.Rest.Request.t(), http_response :: term()) :: t()

Creates a full Crux.Rest.ApiError.t/0 struct, returned / raised by all Crux.Rest functions in case of an API error.