Crux.Rest.CDN (crux_rest v0.3.0-dev)

Functions to generate CDN urls pointing to avatars, icons, etc.

Relevant upstream documentation: Discord Developer Documentation

Types

format_options()
(since 0.1.5)

Specs

format_options() ::
  %{
    optional(:size) => pos_integer(),
    optional(:extension) => String.t(),
    optional(:animated) => boolean()
  }
  | [size: pos_integer(), extension: String.t(), animated: boolean()]

Specifies the file type / extension and size of the resource url to generate.

Notes:

:size has to be any power of two between 16 and 4096

:extension has to be one of "jpg", "jpeg", "png", "webp", "gif"

:animated whether to keep "gif" if the resource is animated regardless of :extension

See the docs of each function for information what sizes / extensions are valid there.

Functions

application_icon(application, options \\ [])
(since 0.3.0)

Specs

application_icon(
  application :: %{icon: String.t() | nil, id: Snowflake.t()},
  format_options()
) :: String.t() | nil

Generates a url to an application icon.

# A map with an icon
iex> %{id: 560524012627820547, icon: "cb861044ce722df3e5a9547a2a012a04"}
...> |> Crux.Rest.CDN.application_icon()
"https://cdn.discordapp.com/app-icons/560524012627820547/cb861044ce722df3e5a9547a2a012a04.webp"

# With format options
iex> %{id: 560524012627820547, icon: "cb861044ce722df3e5a9547a2a012a04"}
...> |> Crux.Rest.CDN.application_icon(extension: "png", size: 2048)
"https://cdn.discordapp.com/app-icons/560524012627820547/cb861044ce722df3e5a9547a2a012a04.png?size=2048"

# A map without an icon
iex> %{id: 560524012627820547, icon: nil}
...> |> Crux.Rest.CDN.application_icon()
nil

base_url()
(since 0.1.5)

Specs

base_url() :: String.t()

Base CDN address.

custom_emoji(emoji)
(since 0.1.5)

Specs

custom_emoji(
  Crux.Structs.Emoji.t()
  | %{id: Crux.Structs.Snowflake.t(), animated: boolean()}
) :: String.t()

Generates the url to an emoji.

This function does not accept any format_options/0.

# A struct
iex> %Crux.Structs.Emoji{id: 438226248293154816, animated: false}
...> |> Crux.Rest.CDN.custom_emoji()
"https://cdn.discordapp.com/emojis/438226248293154816.png"

# A plain map
iex> %{id: 438226248293154816, animated: true}
...> |> Crux.Rest.CDN.custom_emoji()
"https://cdn.discordapp.com/emojis/438226248293154816.gif"

default_user_avatar(user)
(since 0.1.5)

Specs

default_user_avatar(Crux.Structs.User.t() | %{discriminator: String.t()}) ::
  String.t()

Generates a url to the default avatar url of a user.

# A struct
iex> %Crux.Structs.User{discriminator: "0001"}
...> |> Crux.Rest.CDN.default_user_avatar()
"https://cdn.discordapp.com/embed/avatars/1.png"

# A plain map
iex> %{discriminator: "0001"}
...> |> Crux.Rest.CDN.default_user_avatar()
"https://cdn.discordapp.com/embed/avatars/1.png"

group_dm_icon(group_dm, options \\ [])
(since 0.1.5)

Specs

group_dm_icon(
  Crux.Structs.Channel.t()
  | %{id: Crux.Structs.Snowflake.t(), icon: String.t() | nil},
  format_options()
) :: String.t() | nil

Generates the url to group dm channel's icon.

If the group dm channel has no icon, nil will be returned.

# A struct
iex> %Crux.Structs.Channel{id: 354042501201526786, icon: "ecd7839b9eed535f1ae3a545c5d5f3c8"}
...> |> Crux.Rest.CDN.group_dm_icon()
"https://cdn.discordapp.com/channel-icons/354042501201526786/ecd7839b9eed535f1ae3a545c5d5f3c8.webp"

# A plain map
iex> %{id: 354042501201526786, icon: "ecd7839b9eed535f1ae3a545c5d5f3c8"}
...> |> Crux.Rest.CDN.group_dm_icon()
"https://cdn.discordapp.com/channel-icons/354042501201526786/ecd7839b9eed535f1ae3a545c5d5f3c8.webp"

  # With format options
iex> %Crux.Structs.Channel{id: 354042501201526786, icon: "ecd7839b9eed535f1ae3a545c5d5f3c8"}
...> |> Crux.Rest.CDN.group_dm_icon(size: 16, extension: "png")
"https://cdn.discordapp.com/channel-icons/354042501201526786/ecd7839b9eed535f1ae3a545c5d5f3c8.png?size=16"

# Without icon
iex> %{icon: nil}
...> |> Crux.Rest.CDN.group_dm_icon()
nil

guild_banner(guild, options \\ [])
(since 0.2.0)

Specs

guild_banner(
  Crux.Structs.Guild.t()
  | %{id: Crux.Structs.Snowflake.t(), banner: String.t() | nil},
  format_options()
) :: String.t() | nil

Generates a url to a guild banner.

If the guild has no banner, nil will be returned.

The extension "gif" is not valid here.

# A struct
iex> %Crux.Structs.Guild{id: 269508806759809042, banner: "29c1980a3471cb2d5c1208c5196278fb"}
...> |> Crux.Rest.CDN.guild_banner()
"https://cdn.discordapp.com/banners/269508806759809042/29c1980a3471cb2d5c1208c5196278fb.webp"

  # A plain map
iex> %{id: 269508806759809042, banner: "29c1980a3471cb2d5c1208c5196278fb"}
...> |> Crux.Rest.CDN.guild_banner()
"https://cdn.discordapp.com/banners/269508806759809042/29c1980a3471cb2d5c1208c5196278fb.webp"

# With format_options
iex> %Crux.Structs.Guild{id: 269508806759809042, banner: "29c1980a3471cb2d5c1208c5196278fb"}
...> |> Crux.Rest.CDN.guild_banner(size: 16, extension: "png")
"https://cdn.discordapp.com/banners/269508806759809042/29c1980a3471cb2d5c1208c5196278fb.png?size=16"

# Without banner
iex> %Crux.Structs.Guild{id: 269508806759809042, banner: nil}
...> |> Crux.Rest.CDN.guild_banner()
nil

guild_discovery_splash(guild, options \\ [])
(since 0.3.0)

Specs

guild_discovery_splash(
  Crux.Structs.Guild.t()
  | %{id: Crux.Structs.Snowflake.t(), splash: String.t() | nil},
  format_options()
) :: String.t() | nil

Generates a url to a guild's discovery splash.

If the guild has no discovery splash, nil will be returned.

The extension "gif" is not valid here.

# A struct
iex> %Crux.Structs.Guild{id: 269508806759809042, splash: "15abb45cf1c59f90ea291185b99ab1dd"}
...> |> Crux.Rest.CDN.guild_discovery_splash()
"https://cdn.discordapp.com/discovery-splashes/269508806759809042/15abb45cf1c59f90ea291185b99ab1dd.webp"

  # A plain map
iex> %{id: 269508806759809042, splash: "15abb45cf1c59f90ea291185b99ab1dd"}
...> |> Crux.Rest.CDN.guild_discovery_splash()
"https://cdn.discordapp.com/discovery-splashes/269508806759809042/15abb45cf1c59f90ea291185b99ab1dd.webp"

# With format_options
iex> %Crux.Structs.Guild{id: 269508806759809042, splash: "15abb45cf1c59f90ea291185b99ab1dd"}
...> |> Crux.Rest.CDN.guild_discovery_splash(size: 16, extension: "png")
"https://cdn.discordapp.com/discovery-splashes/269508806759809042/15abb45cf1c59f90ea291185b99ab1dd.png?size=16"

# Without splash
iex> %Crux.Structs.Guild{id: 269508806759809042, splash: nil}
...> |> Crux.Rest.CDN.guild_discovery_splash()
nil

guild_icon(guild, options \\ [])
(since 0.1.5)

Specs

guild_icon(
  Crux.Structs.Guild.t()
  | %{id: Crux.Structs.Snowflake.t(), icon: String.t() | nil},
  format_options()
) :: String.t() | nil

Generates a url to a guild's icon.

If the guild has no icon, nil will be returned.

# A guild struct
iex> %Crux.Structs.Guild{id: 269508806759809042, icon: "15abb45cf1c59f90ea291185b99ab1dd"}
...> |> Crux.Rest.CDN.guild_icon()
"https://cdn.discordapp.com/icons/269508806759809042/15abb45cf1c59f90ea291185b99ab1dd.webp"

# A plain map
iex> %{id: 269508806759809042, icon: "15abb45cf1c59f90ea291185b99ab1dd"}
...> |> Crux.Rest.CDN.guild_icon()
"https://cdn.discordapp.com/icons/269508806759809042/15abb45cf1c59f90ea291185b99ab1dd.webp"

# With format_options
iex> %Crux.Structs.Guild{id: 269508806759809042, icon: "15abb45cf1c59f90ea291185b99ab1dd"}
...> |> Crux.Rest.CDN.guild_icon(size: 16, extension: "png")
"https://cdn.discordapp.com/icons/269508806759809042/15abb45cf1c59f90ea291185b99ab1dd.png?size=16"

# Without icon
iex> %Crux.Structs.Guild{id: 269508806759809042, icon: nil}
...> |> Crux.Rest.CDN.guild_icon()
nil

guild_splash(guild, options \\ [])
(since 0.1.5)

Specs

guild_splash(
  Crux.Structs.Guild.t()
  | %{id: Crux.Structs.Snowflake.t(), splash: String.t() | nil},
  format_options()
) :: String.t() | nil

Generates a url to a guild's splash.

If the guild has no splash, nil will be returned.

The extension "gif" is not valid here.

# A struct
iex> %Crux.Structs.Guild{id: 269508806759809042, splash: "15abb45cf1c59f90ea291185b99ab1dd"}
...> |> Crux.Rest.CDN.guild_splash()
"https://cdn.discordapp.com/splashes/269508806759809042/15abb45cf1c59f90ea291185b99ab1dd.webp"

  # A plain map
iex> %{id: 269508806759809042, splash: "15abb45cf1c59f90ea291185b99ab1dd"}
...> |> Crux.Rest.CDN.guild_splash()
"https://cdn.discordapp.com/splashes/269508806759809042/15abb45cf1c59f90ea291185b99ab1dd.webp"

# With format_options
iex> %Crux.Structs.Guild{id: 269508806759809042, splash: "15abb45cf1c59f90ea291185b99ab1dd"}
...> |> Crux.Rest.CDN.guild_splash(size: 16, extension: "png")
"https://cdn.discordapp.com/splashes/269508806759809042/15abb45cf1c59f90ea291185b99ab1dd.png?size=16"

# Without splash
iex> %Crux.Structs.Guild{id: 269508806759809042, splash: nil}
...> |> Crux.Rest.CDN.guild_splash()
nil

member_avatar(member, options \\ [])
(since 0.3.0)

Specs

member_avatar(
  Crux.Structs.Member.t()
  | %{
      avatar: String.t() | nil,
      guild_id: Crux.Structs.Snowflake.t(),
      user: Crux.Structs.Snowflake.t()
    },
  format_options()
) :: String.t() | nil

Generates a url to a guild member avatar.

If the member has no guild avatar, this will return nil.

# A struct
iex> %Crux.Structs.Member{user: 218348062828003328, guild_id: 222078108977594368, avatar: "0c4ec84b8d40bcf8574b904731b930b7"}
...> |> Crux.Rest.CDN.member_avatar()
"https://cdn.discordapp.com/guilds/222078108977594368/users/218348062828003328/avatars/0c4ec84b8d40bcf8574b904731b930b7.webp"

# A plain map with an avatar
iex> %{user: 218348062828003328, guild_id: 222078108977594368, avatar: "0c4ec84b8d40bcf8574b904731b930b7"}
...> |> Crux.Rest.CDN.member_avatar()
"https://cdn.discordapp.com/guilds/222078108977594368/users/218348062828003328/avatars/0c4ec84b8d40bcf8574b904731b930b7.webp"

# No avatar
iex> %{user: 218348062828003328, guild_id: 222078108977594368, avatar: nil}
...> |> Crux.Rest.CDN.member_avatar()
nil

# With format options
iex> %{user: 218348062828003328, guild_id: 222078108977594368, avatar: "0c4ec84b8d40bcf8574b904731b930b7"}
...> |> Crux.Rest.CDN.member_avatar(extension: "png", size: 2048)
"https://cdn.discordapp.com/guilds/222078108977594368/users/218348062828003328/avatars/0c4ec84b8d40bcf8574b904731b930b7.png?size=2048"

role_icon(role, options \\ [])
(since 0.3.0)

Specs

role_icon(
  Crux.Structs.Role.t()
  | %{id: Crux.Structs.Snowflake.t(), icon: String.t() | nil},
  format_options()
) :: String.t() | nil

Generates a url to a role icon.

# A struct
iex> %Crux.Structs.Role{id: 402615751610990610, icon: "536d8aa3468b472500c3d304b2c0c340"}
...> |> Crux.Rest.CDN.role_icon()
"https://cdn.discordapp.com/role-icons/402615751610990610/536d8aa3468b472500c3d304b2c0c340.webp"

# A plain map
iex> %{id: 402615751610990610, icon: "536d8aa3468b472500c3d304b2c0c340"}
...> |> Crux.Rest.CDN.role_icon()
"https://cdn.discordapp.com/role-icons/402615751610990610/536d8aa3468b472500c3d304b2c0c340.webp"

# No icon
iex> %{id: 402615751610990610, icon: nil}
...> |> Crux.Rest.CDN.role_icon()
nil

# Format options
iex> %{id: 402615751610990610, icon: "536d8aa3468b472500c3d304b2c0c340"}
...> |> Crux.Rest.CDN.role_icon(extension: "png", size: 128)
"https://cdn.discordapp.com/role-icons/402615751610990610/536d8aa3468b472500c3d304b2c0c340.png?size=128"

sticker(sticker)
(since 0.3.0)

Specs

sticker(
  Crux.Structs.Sticker.t()
  | %{id: Crux.Structs.Snowflake.t(), format_type: 1..3}
) :: String.t()

Generates a url to a sticker.

The extension is derived from the format type and resizing it is not supported by Discord.

sticker_pack_icon(sticker_pack, options \\ [])
(since 0.3.0)

Specs

sticker_pack_icon(
  %{banner_asset_id: Crux.Structs.Snowflake.t() | nil},
  format_options()
) :: String.t() | nil

Generates a url to a sticker pack icon.

If the pack has no icon, this will return nil.

team_icon(team, options \\ [])
(since 0.3.0)

Specs

team_icon(
  team :: %{icon: String.t() | nil, id: Snowflake.t()},
  format_options()
) :: String.t() | nil

Generates a url to a team icon.

# A map with an icon
iex> %{id: 701002408041512991, icon: "cb861044ce722df3e5a9547a2a012a04"}
...> |> Crux.Rest.CDN.team_icon()
"https://cdn.discordapp.com/team-icons/701002408041512991/cb861044ce722df3e5a9547a2a012a04.webp"

# With format options
iex> %{id: 701002408041512991, icon: "cb861044ce722df3e5a9547a2a012a04"}
...> |> Crux.Rest.CDN.team_icon(extension: "png", size: 2048)
"https://cdn.discordapp.com/team-icons/701002408041512991/cb861044ce722df3e5a9547a2a012a04.png?size=2048"

# A map without an icon
iex> %{id: 701002408041512991, icon: nil}
...> |> Crux.Rest.CDN.team_icon()
nil

user_avatar(user, options \\ [])
(since 0.1.5)

Specs

user_avatar(
  Crux.Structs.User.t()
  | %{
      id: Crux.Structs.Snowflake.t(),
      discriminator: String.t(),
      avatar: String.t() | nil
    },
  format_options()
) :: String.t()

Generates a url to a user avatar.

If the user has no custom avatar, this will return a default one with the extension "png".

The extension defaults to "gif" or "webp" depending on whether the user has an animated avatar.

# A struct with an avatar
iex> %Crux.Structs.User{id: 218348062828003328, avatar: "646a356e237350bf8b8dfde15667dfc4"}
...> |> Crux.Rest.CDN.user_avatar()
"https://cdn.discordapp.com/avatars/218348062828003328/646a356e237350bf8b8dfde15667dfc4.webp"

# A plain map with an avatar
iex> %{id: 218348062828003328, avatar: "646a356e237350bf8b8dfde15667dfc4"}
...> |> Crux.Rest.CDN.user_avatar()
"https://cdn.discordapp.com/avatars/218348062828003328/646a356e237350bf8b8dfde15667dfc4.webp"

# With format options
iex> %Crux.Structs.User{id: 218348062828003328, avatar: "646a356e237350bf8b8dfde15667dfc4"}
...> |> Crux.Rest.CDN.user_avatar(extension: "png", size: 2048)
"https://cdn.discordapp.com/avatars/218348062828003328/646a356e237350bf8b8dfde15667dfc4.png?size=2048"

# A struct without an avatar
iex> %Crux.Structs.User{id: 218348062828003328, avatar: nil, discriminator: "0001"}
...> |> Crux.Rest.CDN.user_avatar()
"https://cdn.discordapp.com/embed/avatars/1.png"

user_banner(user, options \\ [])
(since 0.3.0)

Specs

user_banner(
  Crux.Structs.User.t()
  | %{id: Crux.Structs.Snowflake.t(), banner: String.t() | nil},
  format_options()
) :: String.t() | nil

Generates a url to a user banner.

If the user has no banner, nilwill be returned.

# A struct
iex> %Crux.Structs.User{id: 66564597481480192, banner: "a_27381a7510bf0e6dccd6a5c9ad90cb41"}
...> |> Crux.Rest.CDN.user_banner()
"https://cdn.discordapp.com/banners/66564597481480192/a_27381a7510bf0e6dccd6a5c9ad90cb41.gif"

# A plain map
iex> %{id: 66564597481480192, banner: "a_27381a7510bf0e6dccd6a5c9ad90cb41"}
...> |> Crux.Rest.CDN.user_banner()
"https://cdn.discordapp.com/banners/66564597481480192/a_27381a7510bf0e6dccd6a5c9ad90cb41.gif"

# With format options
iex> %{id: 66564597481480192, banner: "a_27381a7510bf0e6dccd6a5c9ad90cb41"}
...> |> Crux.Rest.CDN.user_banner(size: 200, extension: "webp")
"https://cdn.discordapp.com/banners/66564597481480192/a_27381a7510bf0e6dccd6a5c9ad90cb41.webp?size=200"

# No banner
iex> %{id: 218348062828003328, banner: nil}
...> |> Crux.Rest.CDN.user_banner(size: 200, animated: false)
nil