Crux.Rest.Endpoints (crux_rest v0.3.0-dev)

Endpoints being used by the Crux.Rest module. All functions except base_url/1 are automatically generated depending on the routes Discord documents.

You usually do not need to worry about this module.

Functions

applications()

This function handles the route: /applications.

applications(application_id)

This function handles the route: /applications/:application_id.

applications_commands(application_id)

This function handles the route: /applications/:application_id/commands.

applications_commands(application_id, command_id)

This function handles the route: /applications/:application_id/commands/:command_id.

applications_guilds(application_id)

This function handles the route: /applications/:application_id/guilds.

applications_guilds(application_id, guild_id)

This function handles the route: /applications/:application_id/guilds/:guild_id.

applications_guilds_commands(application_id, guild_id)

This function handles the route: /applications/:application_id/guilds/:guild_id/commands.

applications_guilds_commands(application_id, guild_id, command_id)

This function handles the route: /applications/:application_id/guilds/:guild_id/commands/:command_id.

base_url(version)
(since 0.3.0)

Specs

base_url(version :: integer() | nil) :: String.t()

Base API url, with or without a version.

iex> Crux.Rest.Endpoints.base_url(nil)
"https://discord.com/api"

iex> Crux.Rest.Endpoints.base_url(8)
"https://discord.com/api/v8"

channels()

This function handles the route: /channels.

channels(channel_id)

This function handles the route: /channels/:channel_id.

channels_followers(channel_id)

This function handles the route: /channels/:channel_id/followers.

channels_invites(channel_id)

This function handles the route: /channels/:channel_id/invites.

channels_messages(channel_id)

This function handles the route: /channels/:channel_id/messages.

channels_messages(channel_id, message_id)

This function handles the route: /channels/:channel_id/messages/:message_id.

channels_messages_bulk_delete(channel_id)

This function handles the route: /channels/:channel_id/messages/bulk-delete.

channels_messages_crosspost(channel_id, message_id)

This function handles the route: /channels/:channel_id/messages/:message_id/crosspost.

channels_messages_reactions(channel_id, message_id)

This function handles the route: /channels/:channel_id/messages/:message_id/reactions.

channels_messages_reactions(channel_id, message_id, emoji)

This function handles the route: /channels/:channel_id/messages/:message_id/reactions/:emoji.

channels_messages_reactions(channel_id, message_id, emoji, user_id)

This function handles the route: /channels/:channel_id/messages/:message_id/reactions/:emoji/:user_id.

channels_messages_reactions_me(channel_id, message_id, emoji)

This function handles the route: /channels/:channel_id/messages/:message_id/reactions/:emoji/@me.

channels_permissions(channel_id)

This function handles the route: /channels/:channel_id/permissions.

channels_permissions(channel_id, overwrite_id)

This function handles the route: /channels/:channel_id/permissions/:overwrite_id.

channels_pins(channel_id)

This function handles the route: /channels/:channel_id/pins.

channels_pins(channel_id, message_id)

This function handles the route: /channels/:channel_id/pins/:message_id.

channels_typing(channel_id)

This function handles the route: /channels/:channel_id/typing.

channels_webhooks(channel_id)

This function handles the route: /channels/:channel_id/webhooks.

gateway()

This function handles the route: /gateway.

gateway_bot()

This function handles the route: /gateway/bot.

guilds()

This function handles the route: /guilds.

guilds(guild_id)

This function handles the route: /guilds/:guild_id.

guilds_audit_logs(guild_id)

This function handles the route: /guilds/:guild_id/audit-logs.

guilds_bans(guild_id)

This function handles the route: /guilds/:guild_id/bans.

guilds_bans(guild_id, member_id)

This function handles the route: /guilds/:guild_id/bans/:member_id.

guilds_channels(guild_id)

This function handles the route: /guilds/:guild_id/channels.

guilds_emojis(guild_id)

This function handles the route: /guilds/:guild_id/emojis.

guilds_emojis(guild_id, emoji_id)

This function handles the route: /guilds/:guild_id/emojis/:emoji_id.

guilds_integrations(guild_id)

This function handles the route: /guilds/:guild_id/integrations.

guilds_integrations(guild_id, integration_id)

This function handles the route: /guilds/:guild_id/integrations/:integration_id.

guilds_integrations_sync(guild_id, integration_id)

This function handles the route: /guilds/:guild_id/integrations/:integration_id/sync.

guilds_invites(guild_id)

This function handles the route: /guilds/:guild_id/invites.

guilds_members(guild_id)

This function handles the route: /guilds/:guild_id/members.

guilds_members(guild_id, member_id)

This function handles the route: /guilds/:guild_id/members/:member_id.

guilds_members_me(guild_id)

This function handles the route: /guilds/:guild_id/members/@me.

guilds_members_me_nick(guild_id)

This function handles the route: /guilds/:guild_id/members/@me/nick.

guilds_members_roles(guild_id, member_id)

This function handles the route: /guilds/:guild_id/members/:member_id/roles.

guilds_members_roles(guild_id, member_id, role_id)

This function handles the route: /guilds/:guild_id/members/:member_id/roles/:role_id.

guilds_preview(guild_id)

This function handles the route: /guilds/:guild_id/preview.

guilds_prune(guild_id)

This function handles the route: /guilds/:guild_id/prune.

guilds_regions(guild_id)

This function handles the route: /guilds/:guild_id/regions.

guilds_roles(guild_id)

This function handles the route: /guilds/:guild_id/roles.

guilds_roles(guild_id, role_id)

This function handles the route: /guilds/:guild_id/roles/:role_id.

guilds_templates(guild_id)

This function handles the route: /guilds/:guild_id/templates.

guilds_templates(guild_id, template_code)

This function handles the route: /guilds/:guild_id/templates/:template_code.

guilds_vanity_url(guild_id)

This function handles the route: /guilds/:guild_id/vanity-url.

guilds_webhooks(guild_id)

This function handles the route: /guilds/:guild_id/webhooks.

guilds_widget(guild_id)

This function handles the route: /guilds/:guild_id/widget.

guilds_widget_json(guild_id)

This function handles the route: /guilds/:guild_id/widget.json.

interactions()

This function handles the route: /interactions.

interactions(interaction_id)

This function handles the route: /interactions/:interaction_id.

interactions(interaction_id, interaction_token)

This function handles the route: /interactions/:interaction_id/:interaction_token.

interactions_callback(interaction_id, interaction_token)

This function handles the route: /interactions/:interaction_id/:interaction_token/callback.

invites()

This function handles the route: /invites.

invites(code)

This function handles the route: /invites/:code.

oauth2()

This function handles the route: /oauth2.

oauth2_applications()

This function handles the route: /oauth2/applications.

oauth2_applications_me()

This function handles the route: /oauth2/applications/@me.

oauth2_me()

This function handles the route: /oauth2/@me.

users()

This function handles the route: /users.

users(user_id)

This function handles the route: /users/:user_id.

users_me()

This function handles the route: /users/@me.

users_me_channels()

This function handles the route: /users/@me/channels.

users_me_guilds()

This function handles the route: /users/@me/guilds.

users_me_guilds(guild_id)

This function handles the route: /users/@me/guilds/:guild_id.

voice()

This function handles the route: /voice.

voice_regions()

This function handles the route: /voice/regions.

webhooks()

This function handles the route: /webhooks.

webhooks(webhook_or_application_id)

This function handles the route: /webhooks/:webhook_or_application_id.

webhooks(webhook_or_application_id, token_or_interaction_token)

This function handles the route: /webhooks/:webhook_or_application_id/:token_or_interaction_token.

webhooks_github(webhook_or_application_id, token_or_interaction_token)

This function handles the route: /webhooks/:webhook_or_application_id/:token_or_interaction_token/github.

webhooks_messages(webhook_or_application_id, token_or_interaction_token)

This function handles the route: /webhooks/:webhook_or_application_id/:token_or_interaction_token/messages.

webhooks_messages(webhook_or_application_id, token_or_interaction_token, message_id)

This function handles the route: /webhooks/:webhook_or_application_id/:token_or_interaction_token/messages/:message_id.

webhooks_messages_original(webhook_or_application_id, token_or_interaction_token)

This function handles the route: /webhooks/:webhook_or_application_id/:token_or_interaction_token/messages/@original.

webhooks_slack(webhook_or_application_id, token_or_interaction_token)

This function handles the route: /webhooks/:webhook_or_application_id/:token_or_interaction_token/slack.